Organize wrapper functions under wrapper file
