+++ src/app/api/spotify/create/route.ts
@@ -1,7 +1,7 @@
 import { routeKeyRetriever } from '@/auth/accessKey';
 import { diffBodyParser } from '@/lib/spotify/validators';
 import { sanitize } from 'isomorphic-dompurify';
-import { handlerWithTimeoutAndAuth } from '@/lib/misc/helpers';
+import { handlerWithTimeoutAndAuth } from '@/lib/misc/wrappers';
 import {
 	getPlaylistTracks,
 	createEmptyPlaylist,
