+++ src/app/api/spotify/specific/route.ts
@@ -4,7 +4,7 @@ import {
 	spotAlbumObjectParser,
 	spotPlaylistObjectParser,
 } from '@/lib/spotify/validators';
-import { handlerWithTimeoutAndAuth, threeRetries } from '@/lib/misc/helpers';
+import { handlerWithTimeoutAndAuth, threeRetries } from '@/lib/misc/wrappers';
 import { SPOT_LOGIN_WINDOW, SPOT_URL_BASE } from '@/consts/spotify';
 
 import { NextAuthRequest } from 'next-auth/lib';
